+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - fix: Receipts for parties not showing its name
+- fix: Importing error on i18n
 
 ## v3.4.22 - 2026-04-17